地図情報機能

+++ db/migrate/20211023084412_create_photos.rb
@@ -4,8 +4,8 @@ class CreatePhotos < ActiveRecord::Migration[5.2]
       t.integer :customer_id,              null: false, default: ""
       t.string :image_id,              null: false, default: ""
       t.string :address,              null: false, default: ""
-      t.integer :latitude,              null: false, default: ""
-      t.integer :longitude,              null: false, default: ""
+      t.float :latitude,              null: false, default: ""
+      t.float :longitude,              null: false, default: ""
       t.text :introduction,              null: false, default: ""
       t.timestamps
     end
